requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonify==0.5
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
Werkzeug==2.0.1
